+++ tests/find_iocs_cases/__init__.py
@@ -4,6 +4,7 @@ from .coins import COIN_DATA
 from .cves import CVE_DATA
 from .domains import DOMAIN_DATA
 from .email import EMAIL_DATA
+from .feature__included_ioc_types import individual_included_ioc_types_tests, multiple_included_ioc_types_tests
 from .file_paths import PATH_DATA
 from .hashes import HASH_DATA
 from .ids import ID_DATA
@@ -11,10 +12,13 @@ from .ip_addr import IP_DATA
 from .mac_addr import MAC_DATA
 from .registry_keys import REGISTRY_DATA
 from .tlp_labels import TLP_DATA
+from .urls import URL_DATA
 from .user_agents import UA_DATA
 
 cases = [
     TLP_DATA,
+    individual_included_ioc_types_tests,
+    multiple_included_ioc_types_tests,
     DOMAIN_DATA,
     EMAIL_DATA,
     HASH_DATA,
@@ -27,6 +31,7 @@ cases = [
     ID_DATA,
     COIN_DATA,
     MAC_DATA,
+    URL_DATA,
     UA_DATA,
 ]