+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Removed phone number parsing ([#115](https://github.com/fhightower/ioc-finder/issues/115))
+
 ## [6.0.1] - 2021.06.09
 
 ### Fixed