+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 - Removed phone number parsing ([#155](https://github.com/fhightower/ioc-finder/issues/155))
+- Add support for Python 3.10 ([#188](https://github.com/fhightower/ioc-finder/issues/188))
 
 ## [6.0.1] - 2021.06.09