+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Removed phone number parsing ([#155](https://github.com/fhightower/ioc-finder/issues/155))
 - Dropping support for Python 3.6 ([#187](https://github.com/fhightower/ioc-finder/issues/187))
+- Add support for Python 3.10 ([#188](https://github.com/fhightower/ioc-finder/issues/188))
 
 ## [6.0.1] - 2021.06.09